+++ src/advanced/widgets/admin-header/index.ts
@@ -1,0 +1,1 @@
+export { AdminHeader } from './ui/AdminHeader';
