+++ src/advanced/widgets/cart-header/index.ts
@@ -1,0 +1,1 @@
+export { CartHeader } from './ui/CartHeader';
